small fixes

+++ tools/build-fedora13-deps.bash
@@ -37,7 +37,7 @@ sudo yum install scons libogg-devel python-devel libvorbis-devel openjpeg-devel 
 libcurl-devel expat-devel phonon-devel ogre-devel boost-devel poco-devel \
 pygtk2-devel dbus-devel ccache qt-devel telepathy-farsight-devel libnice-devel \
 bison flex libxml2-devel ois-devel cmake freealut-devel liboil-devel pango-devel \
-wget qt qt4 mercurial unzip libxslt qtscriptbindings \
+wget qt qt4 mercurial unzip libxslt qtscriptbindings freeglut-devel \
 
 if test -f /usr/bin/qmake; then
 	echo qmake exists
@@ -163,19 +163,21 @@ fi
 
 cd $build
 what=xmlrpc-epi
-if test -f $tags/$what-done; then
-    echo $what is done
-else
-    pkgbase=${what}-0.54.2
-    rm -rf $pkgbase
-    zip=../tarballs/$pkgbase.tar.bz2
-    test -f $zip || wget -O $zip http://sourceforge.net/projects/xmlrpc-epi/files/xmlrpc-epi-base/0.54.2/$pkgbase.tar.bz2
-    tar -xjf $zip
-    cd $pkgbase
-    echo yes | ./configure --disable-debug --disable-static
-    make -j $nprocs
-    make install
-    touch $tags/$what-done
+if test -f /usr/local/lib/libxmlrpc-epi-0.0.3.so; then
+	if test -f $tags/$what-done; then
+	    echo $what is done
+	else
+	    pkgbase=${what}-0.54.2
+	    rm -rf $pkgbase
+	    zip=../tarballs/$pkgbase.tar.bz2
+	    test -f $zip || wget -O $zip http://sourceforge.net/projects/xmlrpc-epi/files/xmlrpc-epi-base/0.54.2/$pkgbase.tar.bz2
+	    tar -xjf $zip
+	    cd $pkgbase
+	    echo yes | ./configure --disable-debug --disable-static
+	    make -j $nprocs
+	    sudo make install
+	    touch $tags/$what-done
+	fi
 fi
